Refactored Server Side
API routes and query implementations are now separate
