+++ server/routes/api/contactScan.js
@@ -1,78 +1,14 @@
-const mongoose = require('mongoose');
-const ContactScan = require('../../models/contactScanModel');
-const Authentication = require('../../services/authentication');
-//constants
-//how much time (in minutes) after the last ping should an scanned sheet be reserved for the typer assigned to it.
-const maxReservationDuration = 3;
+const ContactScanUpdater = require('../../services/contactScanUpdater');
+const ContactScanFetcher = require('../../services/contactScanFetcher');
+
 module.exports = (app) => {
-	app.post('/api/contactScan', (req, res, next) => {
-		Authentication.hasRole(req, res, "isOrganizer").then(isUser=>{
-			if(!isUser)
-				return res.json({"error":"missing token"});
-			const scanUrl = req.body.scanUrl;
-			//the cells info as received from the client (i.e. [[{x,y},{x,y},{x,y},{x,y}]...]
-			const cells = req.body.cells;
-			//constructing the cells info as required by the schema (i.e. [{corners:[{x,y},{x,y},{x,y},{x,y}]}...]
-			let cellsObject = []
-			for (let i=0; i<cells.length; i++)
-			{
-				cellsObject[i]={"cells":[]};
-				for(let j=0; j<cells[i].length; j++)
-				{
-					cellsObject[i].cells.push({"corners":cells[i][j]});
-				}
-			}
-			const today = new Date();
-			scanObject={
-				"metadata":{
-					"creationDate": today,
-					"lastUpdate": today,
-					"creatorId": Authentication.getMyId()
-				},
-				"scanUrl":scanUrl,
-				"rows":cellsObject
-			}
-			const newScan = new ContactScan(scanObject);
-			newScan.save(function (err) {
-				if (err){
-					return res.json(err);
-				}
-				else
-					return res.json(req.body);
-			});
-		})
+	app.get('/api/contactScan', (req, res) => {
+		ContactScanFetcher.getContactScan(req, res);
 	});
-	app.get('/api/contactScan', (req, res, next) => {
-		Authentication.hasRole(req, res, "isTyper").then(isUser=>{
-			if(!isUser)
-				return res.json({"error":"missing token"});
-			const typerId = Authentication.getMyId();
-			const now = new Date();
-			const reservationDeadline = new Date(now.getTime()-maxReservationDuration*60000);
-			ContactScan.findOneAndUpdate(
-				{"complete": false, $or:[{"lastPing":null}, {"lastPing":{$lt: reservationDeadline}}, {"typerId": typerId}]},
-				{"$set": {"lastPing": now, "typerId": typerId}},
-				(err, scanData) => {
-				if (err) return res.json({success: false, error: err});
-				if (!scanData)
-					return res.json({"error":"no pending scans are available"});
-				return res.json(scanData);
-			});
-		})
+	app.post('/api/contactScan', (req, res) => {
+		ContactScanUpdater.insertContactScan(req, res);
 	});
-	app.post('/api/contactScan/pingScan', (req, res, next) => {
-		Authentication.hasRole(req, res, "isTyper").then(isUser=>{
-			if(!isUser)
-				return res.json({"error":"missing token"});
-			const scanId = mongoose.Types.ObjectId(req.body.scanId);
-			const now = new Date();
-			ContactScan.update(
-				{"_id": scanId},
-				{"$set": {"lastPing": now}},
-				(err, result) => {
-					return res.json({"err":err, "result":result});
-				}
-			);
-		});
+	app.post('/api/contactScan/pingScan', (req, res) => {
+		ContactScanUpdater.pingScan(req, res);
 	});
 };
